[#4] Implement UI for Splash screen
