+++ app/src/main/java/com/kks/nimblesurveyjetpackcompose/ui/presentation/main/MainActivity.kt
@@ -1,0 +1,34 @@
+package com.kks.nimblesurveyjetpackcompose.ui.presentation.main
+
+import android.os.Bundle
+import android.view.WindowManager
+import androidx.activity.ComponentActivity
+import androidx.activity.compose.setContent
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.tooling.preview.Preview
+import com.kks.nimblesurveyjetpackcompose.ui.presentation.NavGraphs
+import com.kks.nimblesurveyjetpackcompose.ui.theme.NimbleSurveyJetpackComposeTheme
+import com.ramcosta.composedestinations.DestinationsNavHost
+
+class MainActivity : ComponentActivity() {
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+        window.setFlags(
+            WindowManager.LayoutParams.FLAG_LAYOUT_NO_LIMITS,
+            WindowManager.LayoutParams.FLAG_LAYOUT_NO_LIMITS
+        )
+        setContent {
+            NimbleSurveyJetpackComposeTheme {
+                DestinationsNavHost(navGraph = NavGraphs.root)
+            }
+        }
+    }
+}
+
+@Preview(showBackground = true)
+@Composable
+fun DefaultPreview() {
+    NimbleSurveyJetpackComposeTheme {
+        DestinationsNavHost(navGraph = NavGraphs.root)
+    }
+}
