+++ app/src/main/res/drawable/ic_logo_white.xml
@@ -1,0 +1,94 @@
+<vector xmlns:android="http://schemas.android.com/apk/res/android"
+    android:width="202dp"
+    android:height="48dp"
+    android:viewportWidth="202"
+    android:viewportHeight="48">
+  <group>
+    <clip-path
+        android:pathData="M0,0h201.407v48h-201.407z"/>
+    <path
+        android:pathData="M170.445,3.33H164.112L164.112,38.4H170.445V3.33Z"
+        android:fillColor="#ffffff"
+        android:fillType="evenOdd"/>
+    <path
+        android:pathData="M181.955,23.555C182.539,19.76 185.169,17.358 188.92,17.358C192.865,17.358 195.448,19.76 195.593,23.555H181.955ZM188.872,12.506C181.078,12.506 175.623,17.791 175.623,25.621C175.623,33.356 180.933,38.64 189.019,38.64C193.597,38.64 197.396,37.007 199.978,34.22L196.568,30.713C194.717,32.587 192.233,33.596 189.554,33.596C185.754,33.596 182.929,31.386 182.102,27.735H201.341C201.974,18.415 198.029,12.506 188.872,12.506Z"
+        android:fillColor="#ffffff"
+        android:fillType="evenOdd"/>
+    <path
+        android:pathData="M54.617,12.458C50.379,12.505 47.116,14.091 45.265,17.502V12.698H38.933V38.4H45.265V25.766C45.265,21.537 47.797,18.415 52.084,18.367C55.688,18.367 57.88,20.577 57.88,24.18V38.4H64.261V22.354C64.261,16.253 60.608,12.458 54.617,12.458Z"
+        android:fillColor="#ffffff"
+        android:fillType="evenOdd"/>
+    <path
+        android:pathData="M70.02,38.399H76.352V12.697H70.02V38.399Z"
+        android:fillColor="#ffffff"
+        android:fillType="evenOdd"/>
+    <path
+        android:pathData="M116.256,12.458C111.677,12.505 108.267,14.428 106.563,18.607C105.344,14.716 102.179,12.458 97.599,12.458C93.459,12.505 90.245,14.044 88.443,17.454V12.698H82.111V38.4H88.443V25.717C88.443,21.537 90.927,18.462 95.066,18.319C98.574,18.319 100.717,20.528 100.717,24.18V38.4H107.098V25.717C107.098,21.537 109.582,18.462 113.674,18.319C117.18,18.319 119.373,20.528 119.373,24.18V38.4H125.753V22.354C125.753,16.253 122.148,12.458 116.256,12.458Z"
+        android:fillColor="#ffffff"
+        android:fillType="evenOdd"/>
+    <path
+        android:pathData="M145.2,33.26C140.864,33.26 137.845,30.089 137.845,25.477C137.845,20.961 140.864,17.742 145.2,17.742C149.486,17.742 152.506,20.961 152.506,25.477C152.506,30.041 149.486,33.26 145.2,33.26ZM146.612,12.505C142.716,12.505 139.695,14.091 137.845,16.926V3.329H131.512V38.4H137.845V34.22C139.744,37.054 142.764,38.64 146.71,38.64C154.113,38.64 158.936,33.404 158.936,25.669C158.936,17.791 153.968,12.505 146.612,12.505Z"
+        android:fillColor="#ffffff"
+        android:fillType="evenOdd"/>
+    <path
+        android:pathData="M70.02,9.575H76.352V3.329H70.02V9.575Z"
+        android:strokeAlpha="0.3"
+        android:fillColor="#ffffff"
+        android:fillType="evenOdd"
+        android:fillAlpha="0.3"/>
+    <path
+        android:pathData="M0,19.201H9.733V9.601H0V19.201Z"
+        android:strokeAlpha="0.5"
+        android:fillColor="#ffffff"
+        android:fillType="evenOdd"
+        android:fillAlpha="0.5"/>
+    <path
+        android:pathData="M9.733,9.6L19.467,19.2H9.733V9.6Z"
+        android:strokeAlpha="0.6"
+        android:fillColor="#ffffff"
+        android:fillType="evenOdd"
+        android:fillAlpha="0.6"/>
+    <path
+        android:pathData="M0,28.801H9.733V19.201H0V28.801Z"
+        android:strokeAlpha="0.4"
+        android:fillColor="#ffffff"
+        android:fillType="evenOdd"
+        android:fillAlpha="0.4"/>
+    <path
+        android:pathData="M0,48H9.733V38.401H0V48Z"
+        android:strokeAlpha="0.3"
+        android:fillColor="#ffffff"
+        android:fillType="evenOdd"
+        android:fillAlpha="0.3"/>
+    <path
+        android:pathData="M9.733,28.801H19.467V19.201H9.733V28.801Z"
+        android:strokeAlpha="0.5"
+        android:fillColor="#ffffff"
+        android:fillType="evenOdd"
+        android:fillAlpha="0.5"/>
+    <path
+        android:pathData="M19.467,28.801H29.2V19.201H19.467V28.801Z"
+        android:strokeAlpha="0.6"
+        android:fillColor="#ffffff"
+        android:fillType="evenOdd"
+        android:fillAlpha="0.6"/>
+    <path
+        android:pathData="M19.467,9.6H29.2V0H19.467V9.6Z"
+        android:strokeAlpha="0.7"
+        android:fillColor="#ffffff"
+        android:fillType="evenOdd"
+        android:fillAlpha="0.7"/>
+    <path
+        android:pathData="M9.733,28.8H19.467V38.4L9.733,28.8Z"
+        android:strokeAlpha="0.4"
+        android:fillColor="#ffffff"
+        android:fillType="evenOdd"
+        android:fillAlpha="0.4"/>
+    <path
+        android:pathData="M19.467,38.401H29.2V28.801H19.467V38.401Z"
+        android:strokeAlpha="0.5"
+        android:fillColor="#ffffff"
+        android:fillType="evenOdd"
+        android:fillAlpha="0.5"/>
+  </group>
+</vector>
