+++ build.gradle
@@ -1,13 +1,19 @@
 buildscript {
     ext {
-        compose_version = '1.1.0-beta01'
+        compose_version = '1.1.0-rc03'
         detekt_version = '1.21.0'
         jacoco_version = '0.8.7'
+        coil_version = '1.4.0'
+        lifecycle_runtime_version = '2.3.1'
+        activity_compose_version = '1.3.1'
+        constraint_compose_version = '1.0.1'
+        compose_destination_version = '1.1.2-beta'
     }
     dependencies {
         classpath 'com.google.gms:google-services:4.3.13'
         classpath "io.gitlab.arturbosch.detekt:detekt-gradle-plugin:$detekt_version"
         classpath "org.jacoco:org.jacoco.core:$jacoco_version"
+        classpath "org.jetbrains.kotlin:kotlin-gradle-plugin:1.6.10"
     }
 }// Top-level build file where you can add configuration options common to all sub-projects/modules.
 plugins {
